moved compile_list_of_regexes to profiles_filtering and moved it's import in openpype.lib earlier
